requirements.txt
streamlit
pymongo
pandas
plotly
requests
python-dotenv
yfinance
bcrypt
beautifulsoup4
fpdf